+++ KVASS/KACWrapper.cs
@@ -389,8 +389,10 @@ namespace KVASSNS
                     AlarmTypeField = KACAlarmType.GetField("TypeOfAlarm");
                     AlarmTimeProperty = KACAlarmType.GetProperty("AlarmTimeUT");
                     AlarmMarginField = KACAlarmType.GetField("AlarmMarginSecs");
-                    AlarmActionField = KACAlarmType.GetField("AlarmAction");
                     RemainingField = KACAlarmType.GetField("Remaining");
+
+                    AlarmActionField = KACAlarmType.GetField("AlarmAction");
+                    ActionActionProperty = KACAlarmType.GetProperty("AlarmActionConvert");
 
                     XferOriginBodyNameField = KACAlarmType.GetField("XferOriginBodyName");
                     //LogFormatted("XFEROrigin:{0}", XferOriginBodyNameField == null);
@@ -501,11 +503,22 @@ namespace KVASSNS
                 /// <summary>
                 /// What should the Alarm Clock do when the alarm fires
                 /// </summary>
+                //public AlarmActionEnum AlarmAction
+                //{
+                //    get { return (AlarmActionEnum)AlarmActionField.GetValue(actualAlarm); }
+                //    set { AlarmActionField.SetValue(actualAlarm, (Int32)value); }
+                //}
+                /// <summary>
+                /// What should the Alarm Clock do when the alarm fires
+                /// </summary>
                 public AlarmActionEnum AlarmAction
                 {
-                    get { return (AlarmActionEnum)AlarmActionField.GetValue(actualAlarm); }
-                    set { AlarmActionField.SetValue(actualAlarm, (Int32)value); }
+                    get { return (AlarmActionEnum)ActionActionProperty.GetValue(actualAlarm); }
+                    set { ActionActionProperty.SetValue(actualAlarm, (Int32)value); }
                 }
+                private PropertyInfo ActionActionProperty;
+
+
 
                 private FieldInfo RemainingField;
                 /// <summary>
@@ -576,7 +589,7 @@ namespace KVASSNS
                 [Description("Message Only-No Affect on warp")] MessageOnly,
                 [Description("Kill Warp Only-No Message")] KillWarpOnly,
                 [Description("Kill Warp and Message")] KillWarp,
-                [Description("Pause Game and Message")] PauseGame,
+                [Description("Pause Game and Message")] PauseGame
             }
 
             public enum TimeEntryPrecisionEnum